editing scorefunctions.py
